Simplify PathFilters
