+++ detekt-api/src/main/kotlin/io/gitlab/arturbosch/detekt/api/internal/PathFilters.kt
@@ -1,13 +1,7 @@
 package io.gitlab.arturbosch.detekt.api.internal
 
-import io.github.detekt.psi.absolutePath
-import io.github.detekt.psi.basePath
-import io.gitlab.arturbosch.detekt.api.Config
-import org.jetbrains.kotlin.psi.KtFile
 import java.nio.file.Path
 import java.nio.file.PathMatcher
-import kotlin.io.path.Path
-import kotlin.io.path.relativeTo
 
 /**
  * Path filters to explicitly include and/or exclude paths for rules.
@@ -34,17 +28,6 @@ class PathFilters internal constructor(
         return !(isIncluded() && !isExcluded())
     }
 
-    /**
-     * Runs [isIgnored] against a [KtFile] based on its relative path, or based on its [absolutePath] if [basePath] is
-     * not set.
-     */
-    fun isIgnored(ktFile: KtFile): Boolean {
-        ktFile.basePath?.let {
-            return isIgnored(Path(".", ktFile.absolutePath().relativeTo(it).toString()))
-        }
-        return isIgnored(ktFile.absolutePath())
-    }
-
     companion object {
         fun of(includes: List<String>, excludes: List<String>): PathFilters? {
             if (includes.isEmpty() && excludes.isEmpty()) {
@@ -62,10 +45,4 @@ class PathFilters internal constructor(
                     .toSet()
             }
     }
-}
-
-fun Config.createPathFilters(): PathFilters? {
-    val includes = valueOrDefault(Config.INCLUDES_KEY, emptyList<String>())
-    val excludes = valueOrDefault(Config.EXCLUDES_KEY, emptyList<String>())
-    return PathFilters.of(includes, excludes)
 }
